+++ docs/guides/changelog.md
@@ -15,6 +15,7 @@ description: |-
 - Add `TLSv1.3` option to `ip_http_tls_version` attribute of `iosxe_system` resource and data source
 - Rename `snooping_vlans` to `snooping_vlans_legacy` and add `snooping_vlans` attribute to `iosxe_dhcp` resource and data source to support versions >= `17.14`
 - Rename `match_route_type_local` to `match_route_type_local_legacy` and add `match_route_type_local` attribute to `iosxe_route_map` resource and data source to support versions >= `17.15`
+- Remove `delete_mode` attribute from `iosxe_snmp_server_user` resource
 
 ## 0.5.10